refactor : CartItem 을 조회하거나 삭제하는 과정은 CartItemRepository 로 옮김
