+++ src/main/java/cart/repository/OrderRepository.java
@@ -69,22 +69,6 @@ public class OrderRepository {
         return orderProductsAfterSave;
     }
 
-    public List<CartItem> findCartItemsByIds(List<Long> cartItemIds) {
-        return cartItemIds.stream()
-                .map(this::findCartItemById)
-                .collect(Collectors.toList());
-    }
-
-    private CartItem findCartItemById(Long id) {
-        CartItem cartItem = cartItemDao.findById(id);
-
-        if (cartItem == null) {
-            throw new IllegalArgumentException("찾는 장바구니가 없습니다.");
-        }
-
-        return cartItem;
-    }
-
     public List<Order> findOrdersByMember(Member member) {
         List<OrderDto> orderDtos = orderDao.findByMemberId(member.getId());
         return createOrdersFromOrderDtos(member, orderDtos);
@@ -136,20 +120,6 @@ public class OrderRepository {
         Optional<CouponDto> couponDto = couponDao.findById(id);
         return couponDto.map(CouponConvertor::dtoToDomain)
                 .orElse(null);
-    }
-
-    public void deleteCartItems(List<Long> cartItemIds) {
-        for (Long cartItemId : cartItemIds) {
-            deleteCartItem(cartItemId);
-        }
-    }
-
-    private void deleteCartItem(final Long cartItemId) {
-        int deleteCount = cartItemDao.deleteById(cartItemId);
-
-        if (deleteCount == 0) {
-            throw new IllegalArgumentException("장바구니가 삭제되지 않았습니다.");
-        }
     }
 
 }
